requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
pydantic==1.10.12
torch==2.5.1
transformers==4.36.2
tokenizers==0.15.2
accelerate==0.25.0
safetensors==0.4.1
numpy==1.24.3
huggingface-hub==0.20.2
datasets==2.13.0
pyarrow==12.0.1 
requests==2.31.0
filelock==3.13.1
packaging==23.2
psutil==5.9.6
google-cloud-aiplatform==1.28.0